refactor: Reorganizes tests to match DomainOverlay changes.
